fixed code style
